Fixed inconsistent usage of UtcTime
+ Remove all UtcTime usage all over the codebase, for simplicity reasons.
+ Added some Authorization requirements for Booking endpoints
+ Added Authorization requirement for admin endpoints.
+ Fixed a bug where user (as dentist) can't cancel a pending appointment.

+++ DentalClinicBookingPlatform/ClinicPlatformDTOs/ReportModels/AdminSumarizedInfo.cs
@@ -22,7 +22,7 @@ namespace ClinicPlatformObjects.ReportModels
         public int TodayTotal { get; set; }
 
         // Money Informations
-        public long TodayMoneyEarned {  get; set; }
+        public decimal TodayMoneyEarned {  get; set; }
 
         // User Informations
         public int TodayUserCreation {  get; set; }
